on-progress: finish sync play, call, danmaku, popmoji, remaining subtitle share

+++ bunga/server/models.py
@@ -11,29 +11,23 @@ from solo.models import SingletonModel
 
 
 class Site(SingletonModel):
-    name = models.CharField(max_length=100, default='My Bunga Server')
+    name = models.CharField(max_length=100, default="My Bunga Server")
 
 
 class AListHost(SingletonModel):
-    site = models.OneToOneField(Site,
-                                on_delete=models.CASCADE,
-                                primary_key=True)
+    site = models.OneToOneField(Site, on_delete=models.CASCADE, primary_key=True)
     host = models.URLField()
 
 
 class IMKey(SingletonModel):
-    site = models.OneToOneField(Site,
-                                on_delete=models.CASCADE,
-                                primary_key=True)
+    site = models.OneToOneField(Site, on_delete=models.CASCADE, primary_key=True)
     tencent_app_id = models.CharField(max_length=100)
     tencent_app_key = models.CharField(max_length=100)
     tencent_admin_name = models.CharField(max_length=100)
 
 
 class VoiceKey(SingletonModel):
-    site = models.OneToOneField(Site,
-                                on_delete=models.CASCADE,
-                                primary_key=True)
+    site = models.OneToOneField(Site, on_delete=models.CASCADE, primary_key=True)
     agora_key = models.CharField(max_length=100)
     agora_certification = models.CharField(max_length=100)
 
@@ -54,33 +48,28 @@ def create_channel_profile(sender, instance, created, **kwargs):
 
         content_type = ContentType.objects.get_for_model(Channel)
         Permission.objects.create(
-            codename=f'channel_{instance.channel_id}',
-            name=f'Actions in channel {instance.channel_id}',
+            codename=f"channel_{instance.channel_id}",
+            name=f"Actions in channel {instance.channel_id}",
             content_type=content_type,
         )
 
 
 @receiver(models.signals.post_delete, sender=Channel)
 def delete_channel_profile(sender, instance, **kwargs):
-    Permission.objects.filter(
-        codename=f'channel_{instance.channel_id}').delete()
+    Permission.objects.filter(codename=f"channel_{instance.channel_id}").delete()
 
 
 class AListAccount(models.Model):
-    channel = models.OneToOneField(Channel,
-                                   on_delete=models.CASCADE,
-                                   primary_key=True)
-    username = models.CharField(max_length=100, default='')
-    password = models.CharField(max_length=100, default='')
+    channel = models.OneToOneField(Channel, on_delete=models.CASCADE, primary_key=True)
+    username = models.CharField(max_length=100, default="")
+    password = models.CharField(max_length=100, default="")
 
 
 class BilibiliAccount(models.Model):
-    channel = models.OneToOneField(Channel,
-                                   on_delete=models.CASCADE,
-                                   primary_key=True)
-    sess = models.CharField(max_length=500, default='')
-    bili_jct = models.CharField(max_length=200, default='')
-    refresh_token = models.CharField(max_length=200, default='')
+    channel = models.OneToOneField(Channel, on_delete=models.CASCADE, primary_key=True)
+    sess = models.CharField(max_length=500, default="")
+    bili_jct = models.CharField(max_length=200, default="")
+    refresh_token = models.CharField(max_length=200, default="")
 
 
 class VideoRecord(models.Model):
@@ -100,19 +89,19 @@ class VideoRecord(models.Model):
         return f'[{self.channel.channel_id}] {self.title} ({self.updated_at.strftime("%Y-%m-%d %H:%M:%S")})'
 
     class Meta:
-        unique_together = ('channel', 'record_id')
-        ordering = ('channel', '-updated_at')
+        unique_together = ("channel", "record_id")
+        ordering = ("channel", "-updated_at")
 
 
 class Subtitle(models.Model):
     record = models.OneToOneField(
         VideoRecord,
         on_delete=models.CASCADE,
-        related_name='subtitle',
+        related_name="subtitle",
     )
     uploader = models.ForeignKey(
         settings.AUTH_USER_MODEL,
         on_delete=models.CASCADE,
     )
     name = models.CharField(max_length=200)
-    file = models.FileField(upload_to='subtitles/')
+    file = models.FileField(upload_to="subtitles/")
